ablation/runs/requirements.txt
transformers==4.49.0
datasets
peft==0.13.0
sentencepiece
sacrebleu
unbabel-comet
bitsandbytes
accelerate